Refactor ProfileResults
