+++ packages/web/src/pages/search-page-v2/SearchResults.tsx
@@ -4,14 +4,16 @@ import { TrackResultsPage } from './search-results/TrackResults'
 import { AllResults } from './search-results/AllResults'
 import { SearchKind } from '@audius/common/store'
 
+import { ProfileResultsPage } from './search-results/ProfileResults'
+
 export const SearchResults = () => {
   const { category } = useParams<{ category: SearchKind }>()
 
   switch (category) {
     case 'all':
       return <AllResults />
-    // case 'users':
-    //   return <ProfileResults />
+    case 'users':
+      return <ProfileResultsPage />
     case 'tracks':
       return <TrackResultsPage />
     // case 'albums':
